Optimizers python interface (#1441)
* python interface for optimizer

* update lora linear config to support python interface

* update python interface

* finished lora python interface

* fix

* fix

* update

* update

* more fixes

* fix

+++ include/flexflow/ops/kernels/lora_linear_kernels.h
@@ -22,7 +22,7 @@ struct LoraLinearWeight {
 struct LoraLinearModelState {
   LoraLinearWeight weights;
   LoraOptimizerConfig const *optimizer_config;
-  double lora_alpha;
+  float lora_alpha;
 };
 
 class LoraLinearMeta : public OpMeta {
